refactor: Parallelized code
